#6 Fix error build on Linux

+++ Projects/Editor/Source/GUI/Controls/CSplitter.cpp
@@ -704,7 +704,7 @@ namespace Skylicht
 				{
 					current += m_colWidth[i];
 
-					if ((DWORD)x >= current && (DWORD)x <= current + m_expanderSize)
+					if (x >= current && x <= current + m_expanderSize)
 					{
 						isHoverH = true;
 						outCol = i;
